refactor: middleware로직을 `middleware.ts`파일로 옮긴다
