+++ packages/web/src/middleware.ts
@@ -1,8 +1,12 @@
-import { NextRequest } from 'next/server';
-import { updateSession } from './utils/middleware.server';
+import { SUPABASE_ANON_KEY, SUPABASE_URL } from '@src/constants';
+import { createServerClient } from '@supabase/ssr';
+import { NextRequest, NextResponse } from 'next/server';
 
 export async function middleware(request: NextRequest) {
-  return await updateSession(request);
+  const headers = setHeader(request);
+
+  await updateSession(request);
+  return NextResponse.next({ headers });
 }
 
 export const config = {
@@ -17,3 +21,46 @@ export const config = {
     '/((?!_next/static|_next/image|favicon.ico|.*\\.(?:svg|png|jpg|jpeg|gif|webp)$).*)',
   ],
 };
+
+function setHeader(request: NextRequest) {
+  const requestHeaders = new Headers(request.headers);
+  requestHeaders.set('pathname', request.nextUrl.pathname);
+  return requestHeaders;
+}
+
+const HAVE_TO_AUTH_PATH_LIST = ['/memos', 'guide'];
+
+async function updateSession(request: NextRequest) {
+  let supabaseResponse = NextResponse.next({
+    request,
+  });
+
+  const supabase = createServerClient(SUPABASE_URL, SUPABASE_ANON_KEY, {
+    cookies: {
+      getAll() {
+        return request.cookies.getAll();
+      },
+      setAll(cookiesToSet) {
+        // 서버 쿠키 설정
+        cookiesToSet.forEach(({ name, value, options }) => request.cookies.set(name, value));
+        supabaseResponse = NextResponse.next({
+          request,
+        });
+        // 브라우저 쿠키 설정
+        cookiesToSet.forEach(({ name, value, options }) => supabaseResponse.cookies.set(name, value, options));
+      },
+    },
+  });
+
+  const {
+    data: { user },
+  } = await supabase.auth.getUser();
+
+  if (!user && HAVE_TO_AUTH_PATH_LIST.includes(request.nextUrl.pathname)) {
+    const url = request.nextUrl.clone();
+    url.pathname = '/login';
+    return NextResponse.redirect(url);
+  }
+
+  return supabaseResponse;
+}
